Implemented static scalar config provider
